requirements.txt
graphene==2.1.8
Django==2.2.24
django-filter==2.1.0
graphene-file-upload==1.2.2
django-cors-headers==3.7.0
django-s3-storage==0.12.4
boto3==1.17.8
psycopg2-binary==2.8.4
PyJWT==1.7.1
gunicorn==19.9.0
base32-crockford==0.3.0
requests==2.22.0
cryptography==3.3.2
sevenbridges-python==0.24.0
django-rq==2.3.2
rq==1.6.1
rq-scheduler==0.10.0
slack_sdk==3.4.0
django-amazon-ses==3.0.2
xlrd==1.2.0
django-fsm==2.7.0
semantic-version==2.8.5
django-redis-cache==3.0.0
sentry-sdk==1.0.0
kf_lib_data_ingest @ git+https://git@github.com/kids-first/kf-lib-data-ingest.git@1.8.3
graphene_django @ git+https://git@github.com/graphql-python/graphene-django.git@d52b18a700d2c1cec4b2de1f673bff14e2d18071
